requirements.txt
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
google-images-download==2.8.0
idna==2.10
importlib-metadata==4.5.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
redis==3.5.3
requests==2.25.1
rq==1.8.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.5
user-agent==0.1.9
wcwidth==0.2.5
zipp==3.4.1
tqdm